+++ changelog.md
@@ -9,3 +9,5 @@ Github [handlers]
     * updated auth and fixed bug from last commit
 * [github_personal_repo_create_v1] (2021-9-24)
     * updated auth
+* [github_organization_repo_create_v2] (2021-9-24)
+    * added team_id parameter